refactor(mcp): move tool registration into mcpTools.ts; expose JWT email
